feat: Get app functional, with results for typing tests shown for one or more paragraphs (need to bug hunt that feature though)
Note: Careful with the paragraphs - HTML removes some whitespaces, and
manipulating the strings can be tricky. Still need to add accuracy
(different mode that does not need to be 100% accurate to finish), and
make the stats popup on completion (maybe), and definitely clear the
textarea when generating new paragraphs or switching between one or
more paragraph modes. Maybe add some signals to the user that
formatting matters (cannot miss newlines)