feat: Enable rangoli to look more pixelated by removing letters, on an new "no-letters" branch of the repo
Note: Need to update UI for this version so that certain colors
change, not the background since that is now obsolete. Was planning
on doing that anyway, and this looks a bit nicer than the letters
version, but the letters version might be better for checking contrast
and color matches.